SDKCONFIG-21 The getMaxPinFailures might no longer be part of the OneginiConfigModel
In a future release we will remove it but to remain backwards compatible with the Android SDK 6.00.01 release we only removed the @Override annotation that will cause compile issues. Depending on whether there are customers using the 6.00.01 release we will decide to remove the flag completely from the SDK configurator.